simplified build script and added optional keep-html argument

+++ build.js
@@ -2,20 +2,23 @@
 // build.js - Build PDF from a Handlebars and JSON file.
 // Public domain
 
-// Required modules
-var fs = require('fs');
-var handlebars = require('handlebars');
+Size = 'Letter';
+Margin = '.5in';
+
+fs = require('fs');
+system = require('system');
+handlebars = require('handlebars');
 
 // Put the data in the template then produce a temporary HTML file.
-var data = require('./data.json');
-var templateSource = fs.read('template.handlebars');
-var template = handlebars.compile(templateSource);
-var html = template(data);
+data = require('./data.json');
+templateSource = fs.read('template.handlebars');
+template = handlebars.compile(templateSource);
+html = template(data);
 fs.write('_resume.html', html, 'w');
 
 // Render the HTML file into a PDF then delete the HTML file and exit.
-var page = require('webpage').create();
-page.paperSize = { format: 'Letter', orientation: 'portrait', border: '.5in' }
+page = require('webpage').create();
+page.paperSize = { format: Size, orientation: 'portrait', border: Margin }
 page.open('_resume.html', function (status) {
   if (status === 'success') {
     page.render('resume.pdf');
@@ -23,6 +26,6 @@ page.open('_resume.html', function (status) {
   } else {
     console.log("Error: Unable to open page");
   }
-  fs.remove('_resume.html');
+  if (system.args[1] !== '--keep-html') fs.remove('_resume.html');
   phantom.exit();
 });
